+++ src/component.js
@@ -5,6 +5,7 @@ import {
     syncWithObservedComponent,
     updateOnMutation,
 } from './utils'
+import { AlpineMagicHelpers } from './instance'
 
 const AlpineComponentMagicMethod = {
     start() {
@@ -42,6 +43,7 @@ const AlpineComponentMagicMethod = {
     },
 }
 
+window.AlpineMagicHelpers = AlpineMagicHelpers
 const alpine = window.deferLoadingAlpine || ((alpine) => alpine())
 
 window.deferLoadingAlpine = function (callback) {