+++ src/fetch.js
@@ -1,5 +1,6 @@
 import axios from 'axios'
 import { checkForAlpine } from './utils'
+import { AlpineMagicHelpers } from './instance'
 
 const AlpineFetchMagicMethod = {
     start() {
@@ -21,6 +22,7 @@ const AlpineFetchMagicMethod = {
     },
 }
 
+window.AlpineMagicHelpers = AlpineMagicHelpers
 const alpine = window.deferLoadingAlpine || ((alpine) => alpine())
 
 window.deferLoadingAlpine = function (callback) {