+++ src/index.js
@@ -4,7 +4,9 @@ import AlpineIntervalMagicMethod from './interval'
 import AlpineRangeMagicMethod from './range'
 import AlpineScrollMagicMethod from './scroll'
 import AlpineTruncateMagicMethod from './truncate'
+import { AlpineMagicHelpers } from './instance'
 
+window.AlpineMagicHelpers = AlpineMagicHelpers
 const alpine = window.deferLoadingAlpine || ((alpine) => alpine())
 
 window.deferLoadingAlpine = function (callback) {