+++ src/instance.js
@@ -1,0 +1,11 @@
+export const AlpineMagicHelpers = {
+    Config: {
+        breakpoints: {
+            xs: 0,
+            sm: 640,
+            md: 768,
+            lg: 1024,
+            xl: 1280,
+        },
+    },
+}