+++ src/range.js
@@ -1,4 +1,5 @@
 import { checkForAlpine } from './utils'
+import { AlpineMagicHelpers } from './instance'
 
 const AlpineRangeMagicMethod = {
     start() {
@@ -26,6 +27,7 @@ const AlpineRangeMagicMethod = {
     },
 }
 
+window.AlpineMagicHelpers = AlpineMagicHelpers
 const alpine = window.deferLoadingAlpine || ((alpine) => alpine())
 
 window.deferLoadingAlpine = function (callback) {