+++ src/truncate.js
@@ -1,4 +1,5 @@
 import { checkForAlpine } from './utils'
+import { AlpineMagicHelpers } from './instance'
 
 const AlpineTruncateMagicMethod = {
     start() {
@@ -42,6 +43,7 @@ const AlpineTruncateMagicMethod = {
     },
 }
 
+window.AlpineMagicHelpers = AlpineMagicHelpers
 const alpine = window.deferLoadingAlpine || ((alpine) => alpine())
 
 window.deferLoadingAlpine = function (callback) {