+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Adds an instance object to be used by various components
+
 ## [0.4.1] - 2020-11-16
 ### Fixed
 - [$truncate]: Fixes an issue where the ellipsis would show even if the source length was shorter.